moved Multi_Detector_UHF to UHFQC detectors
